backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
langgraph
transformers
torch
torchaudio
soundfile
numpy
python-multipart
python-dotenv
httpx
librosa
